call_center/requirements/base.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-simple-history==3.8.0
djangorestframework==3.16.0
kombu==5.5.3
mysqlclient==2.2.7
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
django-tenants
psycopg2-binary
python-decouple>=3.8
django-environ>=0.10.0
python-dotenv==1.1.0
Pillow==10.0.0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
